Refactor AST headers
