+++ include/ast_clone.h
@@ -13,7 +13,7 @@
 #ifndef VC_AST_CLONE_H
 #define VC_AST_CLONE_H
 
-#include "ast.h"
+#include "ast_expr.h"
 
 /* Recursively clone an expression tree. Returns NULL on allocation failure. */
 expr_t *clone_expr(const expr_t *expr);
