+++ include/semantic_expr.h
@@ -10,7 +10,7 @@
 #ifndef VC_SEMANTIC_EXPR_H
 #define VC_SEMANTIC_EXPR_H
 
-#include "ast.h"
+#include "ast_expr.h"
 #include "ir_core.h"
 #include "symtable.h"
