+++ include/semantic_stmt.h
@@ -10,7 +10,7 @@
 #ifndef VC_SEMANTIC_STMT_H
 #define VC_SEMANTIC_STMT_H
 
-#include "ast.h"
+#include "ast_stmt.h"
 #include "ir_core.h"
 #include "symtable.h"
 #include <stdbool.h>
